ratman/netmod: Async interface for netmods
This uses the #[async_trait] macro from async_trait 0.1, which is
currently the recommended way to do async fn's in traits. However it is
not optimal for performance since it requires indirection and we can
hope that this requirement will be removed eventually.

The poll() and listen() functions are replaced with a single function,
next(), which uses a poll-based interface on an &mut of the netmod,
meaning that only one future from next() can be out at once.