ratman: fix DriverMap
This solution uses a double-locking structure with an RwLock for the
map itself and a Mutex for each Endpoint. This allows us to avoid
concurrent mutable access to the map and the endpoints without taking a
huge performance hit by serializing accesses to the map, since only
.read() access is required in order to lock the mutexes of the endpoints.
That is, only one thread can modify the map at a time (adding or
removing endpoints), and only one thread can access each endpoint at a
time, but all the endpoints can be accessed simultaneously.
This required implementing a custom reference type, DriverRef, which
allows the use of MutexGuard<Box<_>> with a proper lifetime constraint,
without making the caller reliant on the Mutex implementation (so we
can easily change this out if there is a performance issue).