netmod-mem: updating netmod endpoint interface (2/4)
Some changes were required to make this work: for one, the stdlib
channels aren't sync, meaning this crate now relies on the crossbeam
channel implementation. The polling code needed to be changed slightly
to work with the async lock wait while polling for a value to be
returned.  In the future it might be worth it to make the whole
backend work with async-std.

The tests were also updated to no longer create a mutable reference
when linking and sending.